Add SpinningUp examples
Installation is at the beginning of the file. Spinning Up is a library that makes it easier to develop deeper RL models by having more stable and simple implementations. Moreover, SpinningUp is more in-depth, which makes it easier for other researchers to have a better sense of what's going on. For more info, visit spinningup.openai.com. This TD3 uses a custom Long / Short Term Memory Network, which is a branch of the RNN. This neural network will work better for time-series data. A Replay buffer and target network are used for regular updates, alongside explicit loss and optimizer functions. There is an issue with the shaping of the input for the LSTM, but I think someone who understands fishing-v1 better can fix this.